// src/app/@api/model/viewInquiryMainDataDto.ts
/**
 * apis backend
 * Ahmed Abdelrhman & Mohamed sabri
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { InquiryVersionIdDto } from './inquiryVersionIdDto';


export interface ViewInquiryMainDataDto { 
    creationDate?: Date;
    dealId?: number;
    dealNumber?: number;
    inquiryId?: number;
    itemId?: number;
    itemName?: string;
    replies?: Array<InquiryVersionIdDto>;
    status?: ViewInquiryMainDataDto.StatusEnum;
    supplierName?: string;
}
export namespace ViewInquiryMainDataDto {
    export type StatusEnum = 'ACCEPTED' | 'CANCELED' | 'CREATED' | 'SUPPLIER_DRAFTED' | 'SUPPLIER_MODIFIED' | 'USER_DRAFTED' | 'USER_MODIFIED';
    export const StatusEnum = {
        ACCEPTED: 'ACCEPTED' as StatusEnum,
        CANCELED: 'CANCELED' as StatusEnum,
        CREATED: 'CREATED' as StatusEnum,
        SUPPLIERDRAFTED: 'SUPPLIER_DRAFTED' as StatusEnum,
        SUPPLIERMODIFIED: 'SUPPLIER_MODIFIED' as StatusEnum,
        USERDRAFTED: 'USER_DRAFTED' as StatusEnum,
        USERMODIFIED: 'USER_MODIFIED' as StatusEnum
    };
}